cont : n_tokens -> int64_t

+++ tools/server/server-task.h
@@ -557,7 +557,7 @@ struct server_prompt_checkpoint {
     llama_pos pos_min;
     llama_pos pos_max;
 
-    int n_tokens;
+    int64_t n_tokens;
 
     std::vector<uint8_t> data;
